// src/api/models/update-password-model.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Online Resume Builder API
 * An API for Online Resume Builder
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface UpdatePasswordModel
 */
export interface UpdatePasswordModel {
    /**
     * 
     * @type {string}
     * @memberof UpdatePasswordModel
     */
    oldPassword?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpdatePasswordModel
     */
    newPassword?: string | null;
}