// src/api/models/user-vm.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Online Resume Builder API
 * An API for Online Resume Builder
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface UserVM
 */
export interface UserVM {
    /**
     * 
     * @type {string}
     * @memberof UserVM
     */
    id?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UserVM
     */
    firstName: string;
    /**
     * 
     * @type {string}
     * @memberof UserVM
     */
    lastName: string;
    /**
     * 
     * @type {string}
     * @memberof UserVM
     */
    email: string;
}